combine rollup configs
