+++ package.json
@@ -16,12 +16,8 @@
     "codecov": "codecov",
     "precodecov": "npm run coverage",
     "lint": "eslint src test/*.js",
-    "build": "npm run build:main && npm run build:shared && npm run build:ssr",
-    "build:main": "node src/shared/_build.js && rollup -c rollup/rollup.config.main.js",
-    "build:shared": "rollup -c rollup/rollup.config.shared.js",
-    "build:ssr": "rollup -c rollup/rollup.config.ssr.js",
-    "dev": "node src/shared/_build.js && rollup -c rollup/rollup.config.main.js -w",
-    "dev:shared": "rollup -c rollup/rollup.config.shared.js -w",
+    "build": "node src/shared/_build.js && rollup -c",
+    "dev": "node src/shared/_build.js && rollup -c -w",
     "pretest": "npm run build",
     "prepublish": "npm run build && npm run lint",
     "prettier": "prettier --use-tabs --single-quote --trailing-comma es5 --write \"src/**/*.ts\""
@@ -75,7 +71,7 @@
     "nyc": "^10.0.0",
     "prettier": "^1.4.1",
     "reify": "^0.4.4",
-    "rollup": "^0.39.0",
+    "rollup": "^0.43.0",
     "rollup-plugin-buble": "^0.15.0",
     "rollup-plugin-commonjs": "^7.0.0",
     "rollup-plugin-json": "^2.1.0",
