--- rollup/rollup.config.shared.js
@@ -1,5 +1,0 @@
-export default {
-	entry: 'src/shared/index.js',
-	dest: 'shared.js',
-	format: 'es'
-};
